chore: adjust loading, actions sequence, use selector for email code/link flows
